# Fix performance regression results from selection intersection in runtime-filter exploration-phase.